Amend RFC#139 to reduce 40 to 38, omit org/repo
Once this RFC was resolved and Ojas began implementation, it turned out
there were some technical issues with it, not least of which that I had
done the math wrong and 40 characters is too much.  Also, as mentioned
in the PR comments, organization and repository are already allowed to
be 100 characters, so they are removed from consideration here.